refactor: extract tone types and drop funny default
